feat: 32-bit index migration, AOLibre map support, and profile-based storage - Refactored core data structures (GrhInfo, MapData, AssetRegistry) to use 'int' instead of 'short' for indices, supporting GRH IDs beyond 32767. - Implemented AOLibre (Version 136) map loading with 4-byte 'Long' index support. - Fixed binary alignment in MapManager by ensuring particles are always read as 2 bytes (VB6 Integer) to prevent stream desynchronization. - Migrated minimap data storage (minimap.bin) to the 'profiles/' directory for better user data isolation. - Added dynamic memory resizing for AssetRegistry.grhData to handle large, sparse GRH indices. - Improved rendering robustness in Drawn.java with better boundary checks for missing or out-of-range assets. - Updated editor tools (Surface, Npc, Transfer, AutoTiler) to maintain compatibility with the new 32-bit addressing.

+++ src/main/java/org/argentumforge/engine/gui/forms/FMapGenerator.java
@@ -196,8 +196,8 @@ public class FMapGenerator extends Form {
     }
 
     private void applyBiomes(double[][] heightMap, double[][] moistureMap, int width, int height) {
-        Map<TilePos, Short> oldTiles = new HashMap<>();
-        Map<TilePos, Short> newTiles = new HashMap<>();
+        Map<TilePos, Integer> oldTiles = new HashMap<>();
+        Map<TilePos, Integer> newTiles = new HashMap<>();
 
         // Get biome records
         GrhIndexRecord waterRec = categories.get(selectedWaterCategory).getRecords().get(0);
@@ -231,11 +231,11 @@ public class FMapGenerator extends Form {
                 // Apply mosaic pattern
                 int relX = x % biome.getWidth();
                 int relY = y % biome.getHeight();
-                short newGrh = (short) (biome.getGrhIndex() + (relY * biome.getWidth()) + relX);
+                int newGrh = (biome.getGrhIndex() + (relY * biome.getWidth()) + relX);
 
                 int currentGrh = GameData.mapData[x][y].getLayer(1).getGrhIndex();
                 if (currentGrh != newGrh) {
-                    oldTiles.put(new TilePos(x, y), (short) currentGrh);
+                    oldTiles.put(new TilePos(x, y), currentGrh);
                     newTiles.put(new TilePos(x, y), newGrh);
                 }
             }
